Safari-609.1.11.1/LayoutTests/fast/forms/textarea/textarea-state-restore-expected.txt

Confirm nothing is saved and restored:
PASS $("inserted").value is ""
PASS $("saved1").value is ""
PASS $("saved2").value is "aaa\nbbb"
PASS successfullyParsed is true

TEST COMPLETE
